LayoutTests/fast/viewport/viewport-75.html

<head>
    <title>Equals sign for the property starting with "x;". initial-scale is recognized, maximum-scale is not.</title>
    <meta name="viewport" content="width=10 x; maximum-scale=10 initial-scale=10">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
